refactoring some boiler plate code for handling reads
This commit also re-enabled another filter test in RawJsonITest.
