+++ src/test/java/org/usergrid/vx/experimental/RawJsonITest.java
@@ -219,7 +219,7 @@ public class RawJsonITest {
         assertJSONEquals("Failed to apply filter", expectedResponse, actualResponse);
     }
 
-    //@Test
+    @Test
     public void javascriptFilterColumnSlice() throws Exception {
         String insertBeersJSON = loadJSON("insert_beers.json");
         final CountDownLatch doneSignal = new CountDownLatch(1);
